rename novaxr
