+++ csharp-package/brainflow/brainflow/board_controller_library.cs
@@ -49,7 +49,7 @@ namespace brainflow
         CYTON_BOARD = 0,
         GANGLION_BOARD = 1,
         CYTON_DAISY_BOARD = 2,
-        NOVAXR_BOARD = 3,
+        AURAXR_BOARD = 3,
         GANGLION_WIFI_BOARD = 4,
         CYTON_WIFI_BOARD = 5,
         CYTON_DAISY_WIFI_BOARD = 6,
