+++ emulator/brainflow_emulator/auraxr_udp.py
@@ -19,14 +19,35 @@ class State (enum.Enum):
 class Message (enum.Enum):
     start_stream = b'b'
     stop_stream = b's'
-    ack_values = (b'd', b'~6', b'~5')
+    ack_values = (b'd', b'~6')
     ack_from_device = b'A'
     temp_ack_from_host = b'a' # maybe will be removed later
 
 
-class NovaXREmulator (object):
+def test_socket (cmd_list):
+    logging.info ('Running %s' % ' '.join ([str (x) for x in cmd_list]))
+    process = subprocess.Popen (cmd_list, stdout = subprocess.PIPE, stderr = subprocess.PIPE)
+    stdout, stderr = process.communicate ()
+
+    log_multilines (logging.info, stdout)
+    log_multilines (logging.info, stderr)
+
+    if process.returncode != 0:
+        raise TestFailureError ('Test failed with exit code %s' % str (process.returncode), process.returncode)
+
+    return stdout, stderr
+
+
+def run_socket_server ():
+    novaxr_thread = AuraXREmulator ()
+    novaxr_thread.start ()
+    return novaxr_thread
+
+
+class AuraXREmulator (threading.Thread):
 
     def __init__ (self):
+        threading.Thread.__init__ (self)
         self.local_ip = '127.0.0.1'
         self.local_port = 2390
         self.server_socket = socket.socket (family = socket.AF_INET, type = socket.SOCK_DGRAM)
@@ -35,18 +56,18 @@ class NovaXREmulator (object):
         self.state = State.wait.value
         self.addr = None
         self.package_num = 0
-        self.transaction_size = 19
         self.package_size = 72
+        self.keep_alive = True
 
     def run (self):
-        while True:
+        while self.keep_alive:
             try:
                 msg, self.addr = self.server_socket.recvfrom (128)
                 if msg == Message.start_stream.value:
                     self.state = State.stream.value
                 elif msg == Message.stop_stream.value:
                     self.state = State.wait.value
-                elif msg in Message.ack_values.value or msg.decode ('utf-8').startswith ('x'):
+                elif msg in Message.ack_values.value:
                     self.server_socket.sendto (Message.ack_from_device.value, self.addr)
                 elif msg == Message.temp_ack_from_host.value:
                     pass # just remove it from logs
@@ -56,50 +77,35 @@ class NovaXREmulator (object):
                         logging.warn ('received unexpected string %s', str (msg))
             except socket.timeout:
                 logging.debug ('timeout for recv')
+            except Exception:
+                break
 
             if self.state == State.stream.value:
-                transaction = list ()
-                for _ in range (self.transaction_size):
-                    single_package = list ()
-                    for i in range (self.package_size):
-                        single_package.append (random.randint (0, 255))
-                    single_package[0] = self.package_num
-                    
-                    timestamp = bytearray (struct.pack ('d', time.time ()))
-                    eda = bytearray (struct.pack ('f', random.random ()))
-                    ppg_red = bytearray (struct.pack ('i', int (random.random () * 5000)))
-                    ppg_ir = bytearray (struct.pack ('i', int (random.random () * 5000)))
-                    battery = bytearray (struct.pack ('i', int (random.random () * 100)))
-
-                    for i in range (64, 72):
-                        single_package[i] = timestamp[i - 64]
-                    for i in range (1, 5):
-                        single_package[i] = eda[i - 1]
-                    for i in range (60, 64):
-                        single_package[i] = ppg_ir[i - 60]
-                    for i in range (56, 60):
-                        single_package[i] = ppg_red[i - 56]
-                    single_package[53] = random.randint (0, 100)
-
+                package = list ()
+                for _ in range (19):
+                    package.append (self.package_num)
                     self.package_num = self.package_num + 1
                     if self.package_num % 256 == 0:
                         self.package_num = 0
-
-                    transaction.append (single_package)
+                    for i in range (1, self.package_size - 8):
+                        package.append (random.randint (0, 255))
+                    timestamp = bytearray (struct.pack ("d", time.time ()))
+                    package.extend (timestamp)
                 try:
-                    package = list ()
-                    for i in range (self.transaction_size):
-                        package.extend (bytes (transaction[i]))
                     self.server_socket.sendto (bytes (package), self.addr)
                 except socket.timeout:
                     logging.info ('timeout for send')
 
 
-def main ():
-    emulator = NovaXREmulator ()
-    emulator.run ()
+def main (cmd_list):
+    if not cmd_list:
+        raise Exception ('No command to execute')
+    server_thread = run_socket_server ()
+    test_socket (cmd_list)
+    server_thread.keep_alive = False
+    server_thread.join ()
 
 
 if __name__=='__main__':
     logging.basicConfig (format = '%(asctime)s %(levelname)-8s %(message)s', level = logging.INFO)
-    main ()
+    main (sys.argv[1:])
