+++ java-package/brainflow/src/main/java/brainflow/BoardIds.java
@@ -14,7 +14,7 @@ public enum BoardIds
     CYTON_BOARD (0),
     GANGLION_BOARD (1),
     CYTON_DAISY_BOARD (2),
-    NOVAXR_BOARD (3),
+    AURAXR_BOARD (3),
     GANGLION_WIFI_BOARD (4),
     CYTON_WIFI_BOARD (5),
     CYTON_DAISY_WIFI_BOARD (6),
