+++ julia-package/brainflow/src/board_shim.jl
@@ -12,7 +12,7 @@ AnyIntType = Union{Int8, Int32, Int64, Int128, Int}
     CYTON_BOARD = 0
     GANGLION_BOARD = 1
     CYTON_DAISY_BOARD = 2
-    NOVAXR_BOARD = 3
+    AURAXR_BOARD = 3
     GANGLION_WIFI_BOARD = 4
     CYTON_WIFI_BOARD = 5
     CYTON_DAISY_WIFI_BOARD = 6
