+++ matlab-package/brainflow/BoardIDs.m
@@ -6,7 +6,7 @@ classdef BoardIDs < int32
         CYTON_BOARD (0)
         GANGLION_BOARD (1)
         CYTON_DAISY_BOARD (2)
-        NOVAXR_BOARD (3)
+        AURAXR_BOARD (3)
         GANGLION_WIFI (4)
         CYTON_WIFI (5)
         CYTON_DAISY_WIFI (6)
