+++ python-package/brainflow/board_shim.py
@@ -24,7 +24,7 @@ class BoardIds (enum.Enum):
     CYTON_BOARD = 0 #:
     GANGLION_BOARD = 1 #:
     CYTON_DAISY_BOARD = 2 #:
-    NOVAXR_BOARD = 3 #:
+    AURAXR_BOARD = 3 #:
     GANGLION_WIFI_BOARD = 4 #:
     CYTON_WIFI_BOARD = 5 #:
     CYTON_DAISY_WIFI_BOARD = 6 #:
